refactor: split scaffold tests, update doc.go and CHANGELOG
- Split scaffold_test.go (761→344 lines) into three files:
  scaffold_test.go (deployment lifecycle), security_test.go
  (symlink/traversal/containment), contract_test.go (asset contracts)
- Update doc.go to describe both agent and command asset categories
- Add Spec: refs to CHANGELOG entries
- Write code-review: passed marker

Addresses code review findings: AD-007 file size threshold,
Guard doc.go/CHANGELOG findings.
